I'm really tired of thinking of the commit message, so i just push things I made...

+++ src/main/java/com/onxshield/invoiceyou/invoicestatement/model/inventory.java
@@ -19,7 +19,7 @@ public class inventory {
     @OneToOne
     @JoinColumn(name = "fk_product_id", referencedColumnName = "product_id", unique = true)
     private product product;
-    private Long availability;
-    private Long buyPrice;
-    private Long sellPrice;
+    private Double availability;
+    private Double buyPrice;
+    private Double sellPrice;
 }
